Change how execute BDD tests
My thoughts about work with elixir was that I always would do umbrellas projects. So I wrote this framework thinking about it. Another wrong assumption that I've made was that I only should work with .feature files in web projects. And that's not the case at all... So now I made this change to be more close to what I really works.

In the future, maybe I need to change a litle bit more... but for now it works great!