Add ability to run javascript specs
If a container with jasmine is present, then it is possible to just run tests in javascript. It expect that in the root path exists a spec folder and inside it a support folder with the jasmine.json file. The runner uses the root path to call jasmine.

So it needs an elixir container and a javascript container to be able to run those unit and acceptance tests. A interesting dev environment.